fix: make TTP stat show last week calculation
